Modify how contained_node works (#21)
This changes how we represent contained nodes: We now store them as properties, instead of using an edge to access them. 

Furthermore, this changes the API for higher cardinality: We now consider these as List-valued properties.

Furthermore, this changes the interplay between NewNode and StoredNode: It is now possible to directly initialize a StoredNode with properties of a NewNode, without taking a detour via String. 

Furthermore, fields of NewNode are now mutable (`var` instead of `val`). NewNode is still a case class, though.