Make `graph` a method.
This enables us to create derived CPG classes which populate `graph`
lazy. Why do we need this? Well this enables us to create test fixtures
where the builder and the cpg are the same object.
E.g.:
val cpg = code("some code").moreCode("more code")
cpg.typeDecl.l

This becomes valuable in the scalatest setup we often times use where
the cpg in a `should` block is only supposed to be created only if the
test is actually executed (the cpg is accessed).