replace property based mixins with node-specific property accessors
Defining properties globally was a non-optimal decision.
Cardinality and comment are specific to the node where they're used.
There's a few areas where properties are reused with completely different semantics, e.g. fullName.
Even the name and type don't necessarily correlate, i.e. we get into trouble if we introduce a property with a different cardinality or type that has the same name.